+++ src/App.vue
@@ -11,6 +11,7 @@ import TeamSideNav from '@/components/Nav/TeamSideNav'
 import WorkQueueBanner from '@/components/WorkQueueBanner'
 import { eventsMixin } from '@/mixins/eventsMixin'
 import debounce from 'lodash.debounce'
+import VSnackbars from '@/components/Snackbars/Snackbars'
 
 const SERVER_KEY = `${process.env.VUE_APP_RELEASE_TIMESTAMP}_server_url`
 
@@ -46,6 +47,7 @@ export default {
     Footer,
     GlobalSearch,
     TeamSideNav,
+    VSnackbars,
     WorkQueueBanner
   },
   mixins: [eventsMixin],
@@ -471,6 +473,8 @@ export default {
       :timeout="12000"
       :alert-copy="getAlert.alertCopy"
     />
+
+    <VSnackbars />
   </v-app>
 </template>