+++ src/store/alert/index.js
@@ -1,3 +1,5 @@
+import uniqueId from 'lodash.uniqueid'
+
 const state = {
   alert: {
     alertShow: false,
@@ -6,12 +8,20 @@ const state = {
     alertLink: null,
     linkText: '',
     alertCopy: ''
-  }
+  },
+  notifications: [],
+  notificationTimeouts: {}
 }
 
 const getters = {
   getAlert(state) {
     return state.alert
+  },
+  notifications(state) {
+    return state.notifications
+  },
+  notificationTimeouts(state) {
+    return state.notificationTimeouts
   }
 }
 
@@ -28,6 +38,13 @@ const mutations = {
       linkText: '',
       alertCopy: ''
     }
+  },
+  setNotifications(state, notifications = []) {
+    state.notifications = notifications
+  },
+  setNotificationTimeout(state, { id, timeoutId }) {
+    clearTimeout(state['notificationTimeouts'][id])
+    state.notificationTimeouts[id] = timeoutId
   }
 }
 
@@ -35,6 +52,54 @@ const actions = {
   setAlert({ commit }, alert, timeout = 6000) {
     commit('setAl', alert)
     setTimeout(commit, timeout, 'setEmpty')
+  },
+  addNotification({ getters, commit, dispatch }, notification) {
+    const id = uniqueId('notification-')
+    const notifications = [...getters['notifications']]
+    notifications.push({ id: id, ...notification })
+
+    commit('setNotifications', notifications)
+
+    if (notification?.timeout) {
+      const timeoutId = setTimeout(() => {
+        dispatch('dismissNotification', { id })
+      }, notification.timeout)
+
+      commit('setNotificationTimeout', { id, timeoutId })
+    }
+
+    return id
+  },
+  dismissNotification({ getters, commit }, { id }) {
+    const notifications = [...getters['notifications']]
+    const index = notifications.findIndex(n => n.id == id)
+
+    if (index > -1) {
+      notifications.splice(index, 1)
+
+      commit('setNotifications', notifications)
+    }
+  },
+  updateNotification({ getters, commit, dispatch }, { id, notification }) {
+    const notifications = [...getters['notifications']]
+    const index = notifications.findIndex(n => n.id == id)
+
+    if (index > -1) {
+      notifications[index] = {
+        ...notifications[index],
+        ...notification
+      }
+
+      commit('setNotifications', notifications)
+
+      if (notification.timeout) {
+        const timeoutId = setTimeout(() => {
+          dispatch('dismissNotification', { id })
+        }, notification.timeout)
+
+        commit('setNotificationTimeout', { id: notification.id, timeoutId })
+      }
+    }
   }
 }