+++ src/store/tenant/index.js
@@ -191,8 +191,7 @@ const actions = {
         },
         error(error) {
           throw error
-        },
-        fetchPolicy: 'network-only'
+        }
       })
       await dispatch('getTenants')
     } catch (e) {