+++ Changelog.md
@@ -5,6 +5,21 @@
 ### Features and Improvements
 
 - Add webhook and pause schedule actions to automations - [#818](https://github.com/PrefectHQ/ui/pull/818)
+
+### Bugfixes
+
+- None
+
+## 2021-05-25
+
+### Bugfixes
+
+- Remove bad fetchPolicy on tenant settings mutation - [#858](https://github.com/PrefectHQ/ui/pull/858)
+
+## 2021-05-24
+
+### Features and Improvements
+
 - Remove auth layer from the application and improve support for webkit browsers - [#811](https://github.com/PrefectHQ/ui/pull/811)
 - Remove Server/Cloud switcher - [#844](https://github.com/PrefectHQ/ui/pull/844)
 - Allow configurable base url and relative public assets path - [#849](https://github.com/PrefectHQ/ui/pull/849)